+++ src/E5R.Architecture.Data/Abstractions/IAcquirableStorageWithSelector.cs
@@ -18,8 +18,19 @@ namespace E5R.Architecture.Data.Abstractions
     public interface IAcquirableStorageWithSelector<TDataModel> : IStorageSignature
         where TDataModel : IIdentifiable
     {
+        /// <summary>
+        /// Get all stored items
+        /// </summary>
+        /// <param name="projection">Data projection (Include and Select) for data model</param>
+        /// <returns>List of <see cref="IDataProjection{TDataModel, TSelect}.Select"/></returns>
         IEnumerable<TSelect> GetAll<TSelect>(IDataProjection<TDataModel, TSelect> projection);
 
+        /// <summary>
+        /// Gets the stored items in a limited way
+        /// </summary>
+        /// <param name="limiter">Limitation data</param>
+        /// <param name="projection">Data projection (Include and Select) for data model</param>
+        /// <returns>Paged list of <see cref="IDataProjection{TDataModel, TSelect}.Select"/></returns>
         PaginatedResult<TSelect> LimitedGet<TSelect>(IDataLimiter<TDataModel> limiter,
             IDataProjection<TDataModel, TSelect> projection);
     }